+++ src/node-prop.js
@@ -4,6 +4,7 @@
  * @return {!Object}
  */
 module.exports = function(key, val) {
+	//@ts-ignore
 	this.node[key] = val
 	return this
 }